+++ src/Muify.Tests/Domain/ComponentIdentifierComparabilityHasCompareToVisitorTests/WhenObserveIsCalled.cs
@@ -17,9 +17,7 @@ public sealed class WhenObserveIsCalled
             {
                 public int CompareTo(global::MooVC.Testing.Mechanics.Car.Locations? other)
                 {
-                    return other is null
-                        ? 1
-                        : other.CompareTo(Location);
+                    return other is null ? 1 : other.CompareTo(Location);
                 }
             }
             """;
@@ -29,7 +27,9 @@ public sealed class WhenObserveIsCalled
         Component wheel = TestData.Single.Units.Value[0].Components[1]
             .WithMetadata(metadata => metadata
                 .WithIdentifier(identifier => identifier
-                    .WithComparability(comparability => comparability.HasCompareTo(false))));
+                    .WithComparability(comparability => comparability
+                        .HasCompareTo(false)
+                        .IsComparable(Presence.Missing))));
 
         Model.Graph.Areas.Area.Units.Unit.Components.Component component = new(TestData.Single.Components, 0, TestData.Single.Model, wheel);
 
@@ -39,7 +39,7 @@ public sealed class WhenObserveIsCalled
         // Assert
         File definition = await Assert.That(result).HasSingleItem();
         _ = await Assert.That(definition.Content).IsEqualTo(expected);
-        _ = await Assert.That(definition.Hint).IsEqualTo($"{wheel.Name}.Identifier.IEquatable.Equals");
+        _ = await Assert.That(definition.Hint).IsEqualTo($"{wheel.Name}.Identifier.IComparable.CompareTo");
     }
 
     [Test]